replace jalib and jtlib with new link and move instructions
jalib and jtlib are merged into a single link instruction that
uses rc as fun3 which can specify a link function and r6 or r7
as the link register, bit 1 indicates r6 or r7. where fun3 is:

- (0b000) _jib_    or _jmp_  or _jump-ib_
- (0b010) _jalib_  or _call_ or _jump-and-link-ib_
- (0b100) _jtlib_  or _ret_  or _jump-to-link-ib_
- (0b110) _jalaib_ or _tail_ or _jump-and-link-add-ib_

the freed up opcode is repurposed to mov a 32-bit sign-extended
constant into the destination register. movib and addib are
renamed to movw and addh respectively for consistency.

- jalib -> link, jtlib -> movh, movib -> movw, addib -> addh

| 04 | link.i64     | op1r_imm6_16 | 00100 | fun3,ib64(uimm6*8) |
| 05 | movh.i64     | op1r_imm6_16 | 00101 | rc,ib32(uimm6*4)   |
| 06 | movw.i64     | op1r_imm6_16 | 00110 | rc,ib64(uimm6*8)   |
| 12 | addh.i64     | op1r_imm6_16 | 01100 | rc,ib32(uimm6*4)   |